refactor(studio): split workflow orm models
