+++ src/contexts/studio/infrastructure/model_base.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from sqlalchemy.orm import DeclarativeBase
+
+__all__ = ["Base"]
+
+
+class Base(DeclarativeBase):
+    pass
