+++ src/contexts/studio/infrastructure/models.py
@@ -4,18 +4,40 @@ from datetime import datetime
 
 from sqlalchemy import (
     DateTime,
-    Float,
     ForeignKey,
     Integer,
     String,
     Text,
     UniqueConstraint,
 )
-from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
+from sqlalchemy.orm import Mapped, mapped_column, relationship
 
+from src.contexts.studio.infrastructure.model_base import Base
+from src.contexts.studio.infrastructure.workflow_models import (
+    Export,
+    Job,
+    JobEvent,
+    Review,
+    ReviewIssue,
+    UsageEvent,
+)
 
-class Base(DeclarativeBase):
-    pass
+__all__ = [
+    "Base",
+    "Document",
+    "DocumentRevision",
+    "Export",
+    "Job",
+    "JobEvent",
+    "Owner",
+    "Project",
+    "ProjectSnapshot",
+    "Review",
+    "ReviewIssue",
+    "SessionRecord",
+    "SnapshotDocument",
+    "UsageEvent",
+]
 
 
 class Owner(Base):
@@ -193,160 +215,4 @@ class SnapshotDocument(Base):
 
     snapshot: Mapped[ProjectSnapshot] = relationship(
         back_populates="snapshot_documents"
-    )
-
-
-class Job(Base):
-    __tablename__ = "jobs"
-
-    id: Mapped[str] = mapped_column(String(36), primary_key=True)
-    project_id: Mapped[str] = mapped_column(
-        ForeignKey("projects.id", ondelete="CASCADE"),
-        nullable=False,
-        index=True,
-    )
-    document_id: Mapped[str | None] = mapped_column(
-        ForeignKey("documents.id", ondelete="CASCADE"),
-        nullable=True,
-    )
-    kind: Mapped[str] = mapped_column(String(24), nullable=False)
-    operation: Mapped[str] = mapped_column(String(80), nullable=False)
-    status: Mapped[str] = mapped_column(String(24), nullable=False, index=True)
-    provider: Mapped[str] = mapped_column(String(48), default="mock", nullable=False)
-    model: Mapped[str] = mapped_column(String(120), default="", nullable=False)
-    request_json: Mapped[str] = mapped_column(Text, default="{}", nullable=False)
-    result_json: Mapped[str] = mapped_column(Text, default="{}", nullable=False)
-    error: Mapped[str | None] = mapped_column(Text, nullable=True)
-    retry_of_job_id: Mapped[str | None] = mapped_column(
-        ForeignKey("jobs.id", ondelete="SET NULL"),
-        nullable=True,
-    )
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), nullable=False
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), nullable=False
-    )
-    started_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
-    finished_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
-
-    events: Mapped[list[JobEvent]] = relationship(
-        back_populates="job",
-        cascade="all, delete-orphan",
-        order_by="JobEvent.created_at",
-    )
-
-
-class JobEvent(Base):
-    __tablename__ = "job_events"
-
-    id: Mapped[str] = mapped_column(String(36), primary_key=True)
-    job_id: Mapped[str] = mapped_column(
-        ForeignKey("jobs.id", ondelete="CASCADE"),
-        nullable=False,
-        index=True,
-    )
-    status: Mapped[str] = mapped_column(String(24), nullable=False)
-    details_json: Mapped[str] = mapped_column(Text, default="{}", nullable=False)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), nullable=False
-    )
-
-    job: Mapped[Job] = relationship(back_populates="events")
-
-
-class Review(Base):
-    __tablename__ = "reviews"
-
-    id: Mapped[str] = mapped_column(String(36), primary_key=True)
-    project_id: Mapped[str] = mapped_column(
-        ForeignKey("projects.id", ondelete="CASCADE"),
-        nullable=False,
-        index=True,
-    )
-    snapshot_id: Mapped[str] = mapped_column(
-        ForeignKey("project_snapshots.id", ondelete="CASCADE"),
-        nullable=False,
-    )
-    provider: Mapped[str] = mapped_column(String(48), nullable=False)
-    model: Mapped[str] = mapped_column(String(120), nullable=False)
-    summary: Mapped[str] = mapped_column(Text, default="", nullable=False)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), nullable=False
-    )
-
-    issues: Mapped[list[ReviewIssue]] = relationship(
-        back_populates="review",
-        cascade="all, delete-orphan",
-        order_by="ReviewIssue.severity, ReviewIssue.code",
-    )
-
-
-class ReviewIssue(Base):
-    __tablename__ = "review_issues"
-
-    id: Mapped[str] = mapped_column(String(36), primary_key=True)
-    review_id: Mapped[str] = mapped_column(
-        ForeignKey("reviews.id", ondelete="CASCADE"),
-        nullable=False,
-        index=True,
-    )
-    document_id: Mapped[str | None] = mapped_column(
-        ForeignKey("documents.id", ondelete="SET NULL"),
-        nullable=True,
-    )
-    severity: Mapped[str] = mapped_column(String(24), nullable=False)
-    code: Mapped[str] = mapped_column(String(80), nullable=False)
-    message: Mapped[str] = mapped_column(Text, nullable=False)
-    suggestion: Mapped[str] = mapped_column(Text, default="", nullable=False)
-    evidence_json: Mapped[str] = mapped_column(Text, default="{}", nullable=False)
-
-    review: Mapped[Review] = relationship(back_populates="issues")
-
-
-class Export(Base):
-    __tablename__ = "exports"
-
-    id: Mapped[str] = mapped_column(String(36), primary_key=True)
-    project_id: Mapped[str] = mapped_column(
-        ForeignKey("projects.id", ondelete="CASCADE"),
-        nullable=False,
-        index=True,
-    )
-    snapshot_id: Mapped[str] = mapped_column(
-        ForeignKey("project_snapshots.id", ondelete="CASCADE"),
-        nullable=False,
-    )
-    format: Mapped[str] = mapped_column(String(24), nullable=False)
-    relative_path: Mapped[str] = mapped_column(Text, nullable=False)
-    size_bytes: Mapped[int] = mapped_column(Integer, nullable=False)
-    checksum_sha256: Mapped[str] = mapped_column(String(64), nullable=False)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), nullable=False
-    )
-
-
-class UsageEvent(Base):
-    __tablename__ = "usage_events"
-
-    id: Mapped[str] = mapped_column(String(36), primary_key=True)
-    project_id: Mapped[str] = mapped_column(
-        ForeignKey("projects.id", ondelete="CASCADE"),
-        nullable=False,
-        index=True,
-    )
-    job_id: Mapped[str | None] = mapped_column(
-        ForeignKey("jobs.id", ondelete="SET NULL"),
-        nullable=True,
-    )
-    provider: Mapped[str] = mapped_column(String(48), nullable=False)
-    model: Mapped[str] = mapped_column(String(120), nullable=False)
-    prompt_tokens: Mapped[int] = mapped_column(Integer, default=0, nullable=False)
-    completion_tokens: Mapped[int] = mapped_column(Integer, default=0, nullable=False)
-    request_evidence_json: Mapped[str] = mapped_column(
-        Text, default="{}", nullable=False
-    )
-    estimated_cost: Mapped[float | None] = mapped_column(Float, nullable=True)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), nullable=False
     )
